Expose some HiGHS callbacks as .NET events.
The kCallbackLogging, kCallbackMipImprovingSolution, and kCallbackMipInterrupt callbacks are
each exposed as its own event in the HighsLpSolver C# class. The event names are as follows:

    kCallbackLogging => LogReceived
    kCallbackMipImprovingSolution => MipImprovingSolutionFound
    kCallbackMipInterrupt => MipInterruptCheck

Clients subscribe and unsubscribe from these events as they would for any other event,
using syntax such as `lp.LogReceived += MyLoggingFunc`. The client does not need to call
Highs_startCallback or Highs_stopCallback. The events are written so that they automatically
call startCallback when the first client subscribes to the event, and automatically calls
stopCallback when the last client unsubscribes from the event.

The two MIP events include a `user_interrupt` flag in their event args. If an event handler
sets this flag, it is passed back to HiGHS via the cbDataIn callback function argument.