Give protected read-only access to Highs instance pointer in C# HighsLpSolver class.
By exposing the Highs instance as a protected property, the C# HighsLpSolver class can
be sub-classed in useful ways. Without access to the Highs instance, subclasses can't
interact with Highs beyond what is already provided by the main HighsLpSolver class.

Examples of things you might want to do with a subclass include exposing parts of the
Highs C interface that aren't currently implemented in the C# wrapper, and better
support for testing.